+++ package.json
@@ -3,17 +3,24 @@
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
     "main": "index.js",
-    "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
+        "start": "tsc && node dist/index.js",
+        "start:dev": "nodemon",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/node": "^18.0.0",
+        "@types/ws": "^8.5.3",
+        "cross-env": "^7.0.3",
+        "dotenv": "^16.0.1",
+        "eslint": "^8.18.0",
+        "nodemon": "^2.0.16",
+        "ts-node": "^10.8.1",
+        "ts-node-dev": "^2.0.0",
+        "typescript": "^4.7.4"
     },
     "dependencies": {
         "jimp": "^0.16.1",
